go_get_var

Looks for and returns element of given name

go_get_var(x, ...)

Arguments

x
Data object. Particularly useful for recursive lists.

...
Names at each successive level from which to dive and return.

Value

Returns the last named element found in the data object.

Examples

## generate nested list
lst <- list(
  I = 1,
  II = list(
    A = 1,
    B = list(
      a = 1,
      b = list(
        v1 = 1,
        v2 = 2
      )
    )
  )
)
lst <- replicate(2, list(lst))

## fetch and return all v2's
get_var(lst, "II", "B", "v2")

#> Error in get_var(lst, "II", "B", "v2"): could not find function "get_var"