not in

A more convenient way to say no lhs in rhs

lhs %>% rhs

Arguments

lhs
Left hand side

rhs
Right hand side

Value

Logical indicating whether lhs are NOT in rhs. For more info on the output vector.

Details

Number of returned logical values is equal to the length or number of columns of LHS (behaves the same as base in function).

Examples

## is "a" in the alphabet?
"a" %nin% letters

#> [1] FALSE

## vector of "elephant" letters?
elephant <- strsplit("elephant", "")[[1]]

## letters a b and c
abc <- c("a", "b", "c")

## are  "a", "b", or "c" not in
abc %nin% elephant

#> [1] FALSE  TRUE  TRUE