SEM fit statistics

Returns data frame of SEM fit statistics

SEM parameter table

sem_fit(x, digits = 3, scaled = TRUE)

par_table(x)

Arguments

x
Object of class lavaan.

digits
Number of digits to round to, Defaults to 3.

scaled
Logical indicating whether to return scaled fit indices if available.

x
Object of class lavaan

Value

Fit statistics tibble.

Tibble of parameter estimates

Examples

## specify model
m1 <- 'e =~ cyl + hp + wt\nmpg ~ e + gear'

## function to normalize data
normalize <- function(x) (x - mean(x, na.rm = TRUE)) / sd(x, na.rm = TRUE)

## fit model
f1 <- lavaan::sem(
  m1,
  data = tibble::as_tibble(dplyr::mutate_if(mtcars, is.numeric, normalize)),
  std.lv = TRUE,
  estimator = "mlr"
)

## view fit stats
sem_fit(f1)

#> Scaled fit statistics

#> # A tibble: 1 x 7
#>       N ChiSq   CFI   TLI RMSEA      RMSEA_ci  SRMR
#>   <dbl> <dbl> <dbl> <dbl> <dbl>         <chr> <dbl>
#> 1     5 40.88 0.735  0.47 0.474 [0.341,0.619] 0.198

## view par estimates
par_table(f1)

#> # A tibble: 17 x 9
#>      lhs    op   rhs        est      se          z         p   ci.lower
#>    <chr> <chr> <chr>      <dbl>   <dbl>      <dbl>     <dbl>      <dbl>
#>  1     e    =~   cyl  8.964e-01 0.05614  1.597e+01 0.000e+00  7.864e-01
#>  2     e    =~    hp  8.334e-01 0.10405  8.010e+00 1.110e-15  6.295e-01
#>  3     e    =~    wt  8.551e-01 0.12995  6.580e+00 4.702e-11  6.004e-01
#>  4   mpg     ~     e -8.980e-01 0.12456 -7.209e+00 5.627e-13 -1.142e+00
#>  5   mpg     ~  gear  7.687e-02 0.07550  1.018e+00 3.086e-01 -7.110e-02
#>  6   cyl    ~~   cyl  1.652e-01 0.05314  3.109e+00 1.877e-03  6.106e-02
#>  7    hp    ~~    hp  2.742e-01 0.13390  2.048e+00 4.056e-02  1.178e-02
#>  8    wt    ~~    wt  2.376e-01 0.08277  2.871e+00 4.097e-03  7.537e-02
#>  9   mpg    ~~   mpg  9.653e-02 0.05686  1.698e+00 8.958e-02 -1.492e-02
#> 10     e    ~~     e  1.000e+00 0.00000         NA        NA  1.000e+00
#> 11  gear    ~~  gear  9.687e-01 0.00000         NA        NA  9.687e-01
#> 12   cyl    ~1        7.121e-16 0.17399  4.093e-15 1.000e+00 -3.410e-01
#> 13    hp    ~1        2.022e-16 0.17399  1.162e-15 1.000e+00 -3.410e-01
#> 14    wt    ~1        3.072e-16 0.17399  1.766e-15 1.000e+00 -3.410e-01
#> 15   mpg    ~1        1.262e-15 0.16798  7.510e-15 1.000e+00 -3.292e-01
#> 16  gear    ~1       -3.469e-18 0.00000         NA        NA -3.469e-18
#> 17     e    ~1        0.000e+00 0.00000         NA        NA  0.000e+00
#> # ... with 1 more variables: ci.upper <dbl>