map_

maps function and returns list

maps function and returns character vector

maps function and returns double vector

maps function and returns integer vector

maps function and returns logical vector

map_(f, ...)

map_chr_(f, ...)

map_dbl_(f, ...)

map_int_(f, ...)

map_lgl_(f, ...)

Arguments

f
Function

...
Other args passed to function, i.e., named parameters specific to f function, and also the MoreArgs argument from mapply.

f
Function

...
Other args passed to function, i.e., named parameters specific to f function, and also the MoreArgs argument from mapply.

f
Function

...
Other args passed to function, i.e., named parameters specific to f function, and also the MoreArgs argument from mapply.

f
Function

...
Other args passed to function, i.e., named parameters specific to f function, and also the MoreArgs argument from mapply.

f
Function

...
Other args passed to function, i.e., named parameters specific to f function, and also the MoreArgs argument from mapply.

Value

List

Vector type character

Vector type double

Vector type integer

Vector type logical.

Details

Yes, this is identical to base Map but it also hardwires USE.NAMES to FALSE.

Examples

map_(as.data.frame, mtcars)

#> [[1]]
#>    dots[[1L]][[1L]]
#> 1              21.0
#> 2              21.0
#> 3              22.8
#> 4              21.4
#> 5              18.7
#> 6              18.1
#> 7              14.3
#> 8              24.4
#> 9              22.8
#> 10             19.2
#> 11             17.8
#> 12             16.4
#> 13             17.3
#> 14             15.2
#> 15             10.4
#> 16             10.4
#> 17             14.7
#> 18             32.4
#> 19             30.4
#> 20             33.9
#> 21             21.5
#> 22             15.5
#> 23             15.2
#> 24             13.3
#> 25             19.2
#> 26             27.3
#> 27             26.0
#> 28             30.4
#> 29             15.8
#> 30             19.7
#> 31             15.0
#> 32             21.4
#> 
#> [[2]]
#>    dots[[1L]][[2L]]
#> 1                 6
#> 2                 6
#> 3                 4
#> 4                 6
#> 5                 8
#> 6                 6
#> 7                 8
#> 8                 4
#> 9                 4
#> 10                6
#> 11                6
#> 12                8
#> 13                8
#> 14                8
#> 15                8
#> 16                8
#> 17                8
#> 18                4
#> 19                4
#> 20                4
#> 21                4
#> 22                8
#> 23                8
#> 24                8
#> 25                8
#> 26                4
#> 27                4
#> 28                4
#> 29                8
#> 30                6
#> 31                8
#> 32                4
#> 
#> [[3]]
#>    dots[[1L]][[3L]]
#> 1             160.0
#> 2             160.0
#> 3             108.0
#> 4             258.0
#> 5             360.0
#> 6             225.0
#> 7             360.0
#> 8             146.7
#> 9             140.8
#> 10            167.6
#> 11            167.6
#> 12            275.8
#> 13            275.8
#> 14            275.8
#> 15            472.0
#> 16            460.0
#> 17            440.0
#> 18             78.7
#> 19             75.7
#> 20             71.1
#> 21            120.1
#> 22            318.0
#> 23            304.0
#> 24            350.0
#> 25            400.0
#> 26             79.0
#> 27            120.3
#> 28             95.1
#> 29            351.0
#> 30            145.0
#> 31            301.0
#> 32            121.0
#> 
#> [[4]]
#>    dots[[1L]][[4L]]
#> 1               110
#> 2               110
#> 3                93
#> 4               110
#> 5               175
#> 6               105
#> 7               245
#> 8                62
#> 9                95
#> 10              123
#> 11              123
#> 12              180
#> 13              180
#> 14              180
#> 15              205
#> 16              215
#> 17              230
#> 18               66
#> 19               52
#> 20               65
#> 21               97
#> 22              150
#> 23              150
#> 24              245
#> 25              175
#> 26               66
#> 27               91
#> 28              113
#> 29              264
#> 30              175
#> 31              335
#> 32              109
#> 
#> [[5]]
#>    dots[[1L]][[5L]]
#> 1              3.90
#> 2              3.90
#> 3              3.85
#> 4              3.08
#> 5              3.15
#> 6              2.76
#> 7              3.21
#> 8              3.69
#> 9              3.92
#> 10             3.92
#> 11             3.92
#> 12             3.07
#> 13             3.07
#> 14             3.07
#> 15             2.93
#> 16             3.00
#> 17             3.23
#> 18             4.08
#> 19             4.93
#> 20             4.22
#> 21             3.70
#> 22             2.76
#> 23             3.15
#> 24             3.73
#> 25             3.08
#> 26             4.08
#> 27             4.43
#> 28             3.77
#> 29             4.22
#> 30             3.62
#> 31             3.54
#> 32             4.11
#> 
#> [[6]]
#>    dots[[1L]][[6L]]
#> 1             2.620
#> 2             2.875
#> 3             2.320
#> 4             3.215
#> 5             3.440
#> 6             3.460
#> 7             3.570
#> 8             3.190
#> 9             3.150
#> 10            3.440
#> 11            3.440
#> 12            4.070
#> 13            3.730
#> 14            3.780
#> 15            5.250
#> 16            5.424
#> 17            5.345
#> 18            2.200
#> 19            1.615
#> 20            1.835
#> 21            2.465
#> 22            3.520
#> 23            3.435
#> 24            3.840
#> 25            3.845
#> 26            1.935
#> 27            2.140
#> 28            1.513
#> 29            3.170
#> 30            2.770
#> 31            3.570
#> 32            2.780
#> 
#> [[7]]
#>    dots[[1L]][[7L]]
#> 1             16.46
#> 2             17.02
#> 3             18.61
#> 4             19.44
#> 5             17.02
#> 6             20.22
#> 7             15.84
#> 8             20.00
#> 9             22.90
#> 10            18.30
#> 11            18.90
#> 12            17.40
#> 13            17.60
#> 14            18.00
#> 15            17.98
#> 16            17.82
#> 17            17.42
#> 18            19.47
#> 19            18.52
#> 20            19.90
#> 21            20.01
#> 22            16.87
#> 23            17.30
#> 24            15.41
#> 25            17.05
#> 26            18.90
#> 27            16.70
#> 28            16.90
#> 29            14.50
#> 30            15.50
#> 31            14.60
#> 32            18.60
#> 
#> [[8]]
#>    dots[[1L]][[8L]]
#> 1                 0
#> 2                 0
#> 3                 1
#> 4                 1
#> 5                 0
#> 6                 1
#> 7                 0
#> 8                 1
#> 9                 1
#> 10                1
#> 11                1
#> 12                0
#> 13                0
#> 14                0
#> 15                0
#> 16                0
#> 17                0
#> 18                1
#> 19                1
#> 20                1
#> 21                1
#> 22                0
#> 23                0
#> 24                0
#> 25                0
#> 26                1
#> 27                0
#> 28                1
#> 29                0
#> 30                0
#> 31                0
#> 32                1
#> 
#> [[9]]
#>    dots[[1L]][[9L]]
#> 1                 1
#> 2                 1
#> 3                 1
#> 4                 0
#> 5                 0
#> 6                 0
#> 7                 0
#> 8                 0
#> 9                 0
#> 10                0
#> 11                0
#> 12                0
#> 13                0
#> 14                0
#> 15                0
#> 16                0
#> 17                0
#> 18                1
#> 19                1
#> 20                1
#> 21                0
#> 22                0
#> 23                0
#> 24                0
#> 25                0
#> 26                1
#> 27                1
#> 28                1
#> 29                1
#> 30                1
#> 31                1
#> 32                1
#> 
#> [[10]]
#>    dots[[1L]][[10L]]
#> 1                  4
#> 2                  4
#> 3                  4
#> 4                  3
#> 5                  3
#> 6                  3
#> 7                  3
#> 8                  4
#> 9                  4
#> 10                 4
#> 11                 4
#> 12                 3
#> 13                 3
#> 14                 3
#> 15                 3
#> 16                 3
#> 17                 3
#> 18                 4
#> 19                 4
#> 20                 4
#> 21                 3
#> 22                 3
#> 23                 3
#> 24                 3
#> 25                 3
#> 26                 4
#> 27                 5
#> 28                 5
#> 29                 5
#> 30                 5
#> 31                 5
#> 32                 4
#> 
#> [[11]]
#>    dots[[1L]][[11L]]
#> 1                  4
#> 2                  4
#> 3                  1
#> 4                  1
#> 5                  2
#> 6                  1
#> 7                  4
#> 8                  2
#> 9                  2
#> 10                 4
#> 11                 4
#> 12                 3
#> 13                 3
#> 14                 3
#> 15                 4
#> 16                 4
#> 17                 4
#> 18                 1
#> 19                 2
#> 20                 1
#> 21                 1
#> 22                 2
#> 23                 2
#> 24                 4
#> 25                 2
#> 26                 1
#> 27                 2
#> 28                 2
#> 29                 4
#> 30                 6
#> 31                 8
#> 32                 2
#> 

x <- c("A", "brown", "car", "drove", "eighty", "four", "gallons", "how")
map_chr_(function(x) substr(x, 1, 1), x)

#> [1] "A" "b" "c" "d" "e" "f" "g" "h"

x <- rnorm(10)
map_dbl_(round, x, 2)

#>  [1] -0.87  0.11 -0.59 -0.33 -0.09 -2.05  0.15 -0.29  0.25 -0.55

x <- rnorm(10, 100, 5)
map_int_(as.integer, x, 2)

#>  [1] 107  96  92  94 105  96 104  99 104  96

map_lgl_(is.numeric, list(rnorm(10), rnorm(10)))

#> [1] TRUE TRUE