readline_

Worry free way to read lines from interactive sessions.

readline_(...)

Arguments

...
Character string or vector to be used as prompt during interactive R session. Ultimately, this function only sends a single string to the user, but it will accept a vector if you're picky about not creating strings of a certain width.

Value

Input entered during interactive session without extra quotes.