transpose a list

Transpose turns a list-of-lists "inside-out"; it turns a pair of lists into a list of pairs, or a list of pairs into pair of lists. For example, if you had a list of length n where each component had values 'a' and 'b', 'transpose()' would make a list with elements 'a' and 'b' that contained lists of length n. It's called transpose because 'x[[1]][[2]]' is equivalent to 'transpose(x)[[2]][[1]]'.

transpose(.l)

Arguments

.l
A list of vectors to zip. The first element is used as the template; you'll get a warning if a sub-list is not the same length as the first element.

.names
For efficiency, 'transpose()' usually inspects the first component of '.l' to determine the structure. Use '.names' if you want to override this default.

Value

A list with indexing transposed compared to '.l'.

Examples

x <- rerun(5, x = runif(1), y = runif(5))
x %>% str()

#> List of 5
#>  $ :List of 2
#>   ..$ x: num 0.7
#>   ..$ y: num [1:5] 0.22 0.728 0.217 0.456 0.333
#>  $ :List of 2
#>   ..$ x: num 0.568
#>   ..$ y: num [1:5] 0.252 0.464 0.918 0.973 0.819
#>  $ :List of 2
#>   ..$ x: num 0.903
#>   ..$ y: num [1:5] 0.581 0.773 0.995 0.711 0.215
#>  $ :List of 2
#>   ..$ x: num 0.292
#>   ..$ y: num [1:5] 0.7218 0.8666 0.2385 0.0045 0.9435
#>  $ :List of 2
#>   ..$ x: num 0.438
#>   ..$ y: num [1:5] 0.751 0.668 0.408 0.351 0.738

x %>% transpose() %>% str()

#> List of 2
#>  $ x:List of 5
#>   ..$ : num 0.7
#>   ..$ : num 0.568
#>   ..$ : num 0.903
#>   ..$ : num 0.292
#>   ..$ : num 0.438
#>  $ y:List of 5
#>   ..$ : num [1:5] 0.22 0.728 0.217 0.456 0.333
#>   ..$ : num [1:5] 0.252 0.464 0.918 0.973 0.819
#>   ..$ : num [1:5] 0.581 0.773 0.995 0.711 0.215
#>   ..$ : num [1:5] 0.7218 0.8666 0.2385 0.0045 0.9435
#>   ..$ : num [1:5] 0.751 0.668 0.408 0.351 0.738

# Back to where we started
x %>% transpose() %>% transpose() %>% str()

#> List of 5
#>  $ :List of 2
#>   ..$ x: num 0.7
#>   ..$ y: num [1:5] 0.22 0.728 0.217 0.456 0.333
#>  $ :List of 2
#>   ..$ x: num 0.568
#>   ..$ y: num [1:5] 0.252 0.464 0.918 0.973 0.819
#>  $ :List of 2
#>   ..$ x: num 0.903
#>   ..$ y: num [1:5] 0.581 0.773 0.995 0.711 0.215
#>  $ :List of 2
#>   ..$ x: num 0.292
#>   ..$ y: num [1:5] 0.7218 0.8666 0.2385 0.0045 0.9435
#>  $ :List of 2
#>   ..$ x: num 0.438
#>   ..$ y: num [1:5] 0.751 0.668 0.408 0.351 0.738

# transpose() is useful in conjunction with safely() & quietly()
x <- list("a", 1, 2)
y <- x %>% map(safely(log))
y %>% str()

#> List of 3
#>  $ :List of 2
#>   ..$ result: NULL
#>   ..$ error :List of 2
#>   .. ..$ message: chr "non-numeric argument to mathematical function"
#>   .. ..$ call   : language log(x = x, base = base)
#>   .. ..- attr(*, "class")= chr [1:3] "simpleError" "error" "condition"
#>  $ :List of 2
#>   ..$ result: num 0
#>   ..$ error : NULL
#>  $ :List of 2
#>   ..$ result: num 0.693
#>   ..$ error : NULL

y %>% transpose() %>% str()

#> List of 2
#>  $ result:List of 3
#>   ..$ : NULL
#>   ..$ : num 0
#>   ..$ : num 0.693
#>  $ error :List of 3
#>   ..$ :List of 2
#>   .. ..$ message: chr "non-numeric argument to mathematical function"
#>   .. ..$ call   : language log(x = x, base = base)
#>   .. ..- attr(*, "class")= chr [1:3] "simpleError" "error" "condition"
#>   ..$ : NULL
#>   ..$ : NULL

# Use simplify_all() to reduce to atomic vectors where possible
x <- list(list(a = 1, b = 2), list(a = 3, b = 4), list(a = 5, b = 6))
x %>% transpose()

#> $a
#> $a[[1]]
#> [1] 1
#> 
#> $a[[2]]
#> [1] 3
#> 
#> $a[[3]]
#> [1] 5
#> 
#> 
#> $b
#> $b[[1]]
#> [1] 2
#> 
#> $b[[2]]
#> [1] 4
#> 
#> $b[[3]]
#> [1] 6
#> 
#> 

x %>% transpose() %>% simplify_all()

#> $a
#> [1] 1 3 5
#> 
#> $b
#> [1] 2 4 6
#> 

# Provide explicit component names to prevent loss of those that don't
# appear in first component
ll <- list(
  list(x = 1, y = "one"),
  list(z = "deux", x = 2)
)
ll %>% transpose()

#> $x
#> $x[[1]]
#> [1] 1
#> 
#> $x[[2]]
#> [1] 2
#> 
#> 
#> $y
#> $y[[1]]
#> [1] "one"
#> 
#> $y[[2]]
#> NULL
#> 
#> 

nms <- ll %>% map(names) %>% reduce(union)
ll %>% transpose(.names = nms)

#> $x
#> $x[[1]]
#> [1] 1
#> 
#> $x[[2]]
#> [1] 2
#> 
#> 
#> $y
#> $y[[1]]
#> [1] "one"
#> 
#> $y[[2]]
#> NULL
#> 
#> 
#> $z
#> $z[[1]]
#> NULL
#> 
#> $z[[2]]
#> [1] "deux"
#> 
#>