SEM fit statistics

Returns data frame of SEM fit statistics

SEM parameter table

sem_fit(x, digits = 3, scaled = TRUE)

par_table(x)

Arguments

x
Object of class lavaan.

digits
Number of digits to round to, Defaults to 3.

scaled
Logical indicating whether to return scaled fit indices if available.

x
Object of class lavaan

Value

Fit statistics tibble.

Tibble of parameter estimates

Examples

## specify model
m1 <- 'e =~ cyl + hp + wt\nmpg ~ e + gear'

## function to normalize data
normalize <- function(x) (x - mean(x, na.rm = TRUE)) / sd(x, na.rm = TRUE)

## fit model
f1 <- lavaan::sem(
  m1,
  data = tibble::as_tibble(dplyr::mutate_if(mtcars, is.numeric, normalize)),
  std.lv = TRUE,
  estimator = "mlr"
)

## view fit stats
sem_fit(f1)

#> Scaled fit statistics

#> # A tibble: 1 x 7
#>       N  ChiSq   CFI   TLI RMSEA      RMSEA_ci  SRMR
#>   <dbl>  <dbl> <dbl> <dbl> <dbl>         <chr> <dbl>
#> 1     5 40.881 0.735  0.47 0.474 [0.341,0.619] 0.198

## view par estimates
par_table(f1)

#> # A tibble: 17 x 9
#>      lhs    op   rhs           est         se             z            p
#>    <chr> <chr> <chr>         <dbl>      <dbl>         <dbl>        <dbl>
#>  1     e    =~   cyl  8.964025e-01 0.05613879  1.596761e+01 0.000000e+00
#>  2     e    =~    hp  8.333885e-01 0.10404560  8.009839e+00 1.110223e-15
#>  3     e    =~    wt  8.550734e-01 0.12994897  6.580070e+00 4.702261e-11
#>  4   mpg     ~     e -8.980040e-01 0.12456280 -7.209247e+00 5.626610e-13
#>  5   mpg     ~  gear  7.687214e-02 0.07549680  1.018217e+00 3.085747e-01
#>  6   cyl    ~~   cyl  1.652121e-01 0.05313885  3.109064e+00 1.876809e-03
#>  7    hp    ~~    hp  2.742140e-01 0.13389583  2.047965e+00 4.056339e-02
#>  8    wt    ~~    wt  2.375991e-01 0.08277030  2.870584e+00 4.097141e-03
#>  9   mpg    ~~   mpg  9.652980e-02 0.05686186  1.697620e+00 8.957959e-02
#> 10     e    ~~     e  1.000000e+00 0.00000000            NA           NA
#> 11  gear    ~~  gear  9.687500e-01 0.00000000            NA           NA
#> 12   cyl    ~1        7.120990e-16 0.17399264  4.092696e-15 1.000000e+00
#> 13    hp    ~1        2.021814e-16 0.17399264  1.162011e-15 1.000000e+00
#> 14    wt    ~1        3.072168e-16 0.17399264  1.765688e-15 1.000000e+00
#> 15   mpg    ~1        1.261553e-15 0.16797890  7.510185e-15 1.000000e+00
#> 16  gear    ~1       -3.469447e-18 0.00000000            NA           NA
#> 17     e    ~1        0.000000e+00 0.00000000            NA           NA
#> # ... with 2 more variables: ci.lower <dbl>, ci.upper <dbl>